Override, abstract and organise entire Storm library
I was really trying to avoid having to do this, but there's waaaaaay too many circular references and imports in the core Storm library to be able to accurately modify stuff. I assume originally the elements were set up to be able to rendered on their own, but it just doesn't work, and given Storm in our case is used entirely in the Backend only, there's no point.

But, man... this was a metric ton of work to do.